- fixed up some more gcc3.3 errors
git-svn-id: http://svn.ratbox.org/svnroot/trunk@13963 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ adns/query.c
@@ -279,7 +279,7 @@ int adns_submit_reverse_ip6(adns_state ads,
   if(addr->sa_family != AF_INET6) return ENOSYS;
   cp = (const unsigned char *)&(((const struct sockaddr_in6*)addr) -> sin6_addr.s6_addr);
 	lreq = 71 + strlen(zone) + 1;
-  if (lreq > sizeof(shortbuf)) {
+  if ((size_t)lreq > sizeof(shortbuf)) {
     buf= MyMalloc(strlen(zone) + 4*16 + 1);
 #if 0
     if (!buf) return errno;
